SAT-734: Document how to package and install SAT product stream
Add documentation for packaging a SAT release distribution and for
installing that release distribution on a system.

Add these docs to this repository's `README.md`. This may need to go
somewhere else in the future. It probably needs to be adapted into a
customer-facing document.

Test Description:
Examined `README.md` rendered at stash.us.cray.com