SAT-352: Add SAT to cray-product-catalog during install
This adds a step to the install process to run a podman command
that uses the cray-product-catalog-update:0.0.3 container image
to add an entry to the cray-product-catalog Kubernetes config map.

This commit also adds a step to the release process to dynamically
substitute in the SAT version number in docker/index.yaml and in
cray-product-catalog/sat.yaml, the data that will eventually end up
in the config map. This way, the version number only needs to be
maintained in one place.

Test Description: I built this release using release.sh,
and then I installed it on sif. Then, I ran `sat showrev`
and verified the SAT product was in the output.