SAT-743: Use arti docker repositories instead of dtr
Use the docker repositories in arti.dev.cray.com instead of the cray
repository from dtr in preparation for the coming shutdown of dtr.

Once the images are downloaded, move them all into a common "cray"
repository instead of their individual product-specific repos where they
come from in arti. This was per guidance from Zach Crisler in slack in
the #shasta-product-streams channel. Since the repo in the Nexus docker
registry has not changed, scripts which reference these images do not
need to change.

Add logic that determines the arti repository to pull the cray-sat image
from based on the git branch of this repository. The master branch will
pull from `sat-docker-master-local`, and the release branches will pull
from `sat-docker-stable-local`.

Test Description:
Built ISO and installed on thanos. Verified that the new version of the
sat container image was placed in the cray repository in the container
image registry in Nexus.