SAT-743: Version cray-product-catalog-update in one place
Specify the version of the cray-product-catalog-update (abbreviated
CPCU) docker image in a single place in `component_versions.sh`, so that
it can be used in two places: the docker manifest, and the install
script.

This involves substituting a version string in the install.sh script
using sed, so rename this file to `install.sh.in` here, so that we can
tell this file need preprocessing. Add `install.sh` to `.gitignore`.
Ensure the generated `install.sh` has execute permissions. Remove
execute permissions from the `install.sh.in` file because they're not
needed.

Test Description:
Built release distribution locally, copied to thanos-ncn-m001, and
installed. Verified that product catalog was updated as expected.