SAT-936: Use version.sh script instead of .version file
The Jenkins pipeline shared library for building a release distribution
behaves differently when using a `.version` file instead of a version
script. Use a `version.sh` script instead since we prefer that behavior.
This will allow us to have our master branch build with a version that
is suitable for use in a Kubernetes job name, which will allow the
cf-gitea-import K8s job to succeed.

This change still results in only the builds from our release branches,
i.e. branches that start with "release/", getting version number that
are of the form x.y.z. Builds from all other branches will still include
the git commit hash and timestamp, and they will be published to the
shasta-distribution-unstable-local location in arti. This will continue
to make it obvious which builds are true release builds.

Test Description:
Ran `version.sh` locally on this branch with and without uncommitted
changes. Created a temporary `release/sat-2.1` branch and ran
`version.sh` there to verify that it output a plain version number
without the timestamp and git commit hash.

Built a release distribution from this branch and inspected to make sure
the product version specified in the release distribution tar.gz file
and in the loftsman manifest was correct.