SAT-901: Use manifest file for RPMs instead of bloblet location
Create a manifest file for RPMs instead of using the bloblet location.
This has several advantages:

* The product stream explicitly specifies the version of the RPMs it
  delivers, and it can put this information in the cray-product-catalog.
* If the explicitly specified version of any RPM is not available, the
  build will fail. With the bloblet, it was possible to generate a
  release distribution that contained unexpected or incorrect versions.
* There is no longer any need to wait for a separate Jenkins job managed
  by the DST team to populate the bloblet location with RPMs from master
  and release branches.
* There is no longer a circular dependency when creating a new release
  branch. When we used the bloblet location, it was necessary to create
  and push a release branch in order for the release-branch bloblet
  location to be populated, but we must specify that release-branch
  bloblet location in the `release.sh` script in order to generate a
  correct release distribution.

The only known disadvantage of this approach is that it is required to
explicitly update the RPM version in RPM manifest when it has built and
is available. This already must be done for docker images.

Also use hard-coded values for SAT_REPO_TYPE and CAR_REPO If we try to
determine these based on the branch name, it won't work for branches
based on the release branch and PRs targeting the release branch since
the branch name will not be of the form "release/*", but the artifacts
should come from the release location in arti and car.

Bump minor version on master branch to signify divergence from content
in release branch release/sat-2.1.

Test Description:
Built release distribution and examined contents. Installed on baldar
and verified that new package repository was created in Nexus. Ran
through full install procedure to verify that the CFS session ran
successfully and installed a new version of the cray-sat-podman RPM from
the package repository.

(cherry picked from commit cdc73dddb8b75ddc9eb4195b289ccb84c3bc3be0)