SAT-902: Download RPMs from arti instead of car
Download RPMs from arti instead of from car since car is expected to be
shut down later this year. The path in arti unfortunately contains
"dev/master" or "release/sat-2.1", so add some logic to `release.sh`
that will set a variable for the arti subdirectory appropriately based
on whether the `SAT_REPO_TYPE` is "master" or not.

Test Description:
Built a release distribution locally and verified it pulled the correct
cray-sat-podman RPM, put it in the same location as it did when pulling
from car, and correctly generated the repodata.

Will test installation on a system (thanos) by running `install.sh` and
then upgrading to the latest cray-sat-podman RPM on ncn-m001 by running
`zypper up cray-sat-podman` to ensure the RPM is still installable. Will
then run some `sat` commands with the new wrapper script RPM in place
before cleaning up.