CRAYSAT-1108: Product stream version 2.1.14
This commit increments the version of the SAT 2.1 product stream,
to pull in SAT version 3.7.2 and cray-sat-podman version 1.4.8.
This is to pull in a critical fix for CRAYSAT_1106.

Test Description:
I built and installed this release distribution, then verified
that SAT version 3.7.2 was used by running `sat --version`.
Further testing of CRAYSAT_1106 is linked to that Jira.