CRAYSAT-1038: Wait for completion of config import job
This commit adds a `kubectl wait` command into the install script, which
waits for the completion of the sat-config-import-x.y.z Kubernetes job
which has just been launched as a part of deploying the loftsman
manifest. If it fails, the installer prints out a helpful message with
troubleshooting steps.

This commit also changes the output of the install script significantly,
by removing the `set -x` from the start of the script and adding more
informational messages about what is happening at each stage of the
install. To facilitate better info and error messages, functions
exit_with_error and print_stage are added.

In case the extra output is desired, the environment variable
$SAT_INSTALL_DEBUG can be set.

Test Description:
I built a release distribution and installed it on an internal system.
Then, I slightly modified the `kubectl wait` command so that it would
fail, and inspected the resulting error message. I also ran with
$SAT_INSTALL_DEBUG set.