CRAYSAT-1132: Update sat-product-stream to fit schema
This commit changes the format of the data inserted into the
product catalog at install time to conform to the new schema.

Test Description: I built the product stream and used an unstable
version of cray-product-catalog-update that included schema validation
so that I could validate the data going into the product catalog
at install time. I inspected the output to verify the schema
was validated.