CRAYSAT-1171: Integrate cfs-config-util in product stream
This change adds a step to `install.sh` in the SAT release tarball which
calls the cfs-config-util container to add a layer to the active CFS
configuration for the newly installed version. To do this, this change
adds the docker image to the tarball and syncs it to Nexus, then calls
it at the end of the install flow.

Test Description:
Run full installation flow on thanos to exercise the full process.