CRAYSAT-1192: Update cray-product-catalog-update and validate schema
This commit updates the version of cray-product-catalog-update used by
the SAT product stream to a version that supports schema validation and
setting an 'active' flag on install. The image is pulled from a
different repository within arti.dev.cray.com, therefore the release
script is modified slightly to ensure all the images end up in the same
`cray/` repository.

Additionally, this commit modifies the call to
cray-product-catalog-update within the install script, so that the
VALIDATE_SCHEMA variable is passed to the container so that schema
validation actually occurs on install.

Finally, the call to cray-product-catalog-update is updated to include
--user root, which is required to access the Kubernetes configuration
when running this image via podman.

Test Description: I built the release distribution and inspected the
layout of the `docker/` subdirectory to make sure all the images were
under `docker/arti.dev.cray.com/cray/`.

Next, I installed the release distribution and inspected the output
to ensure that cray-product-catalog-update version 1.2.70 was being
used, and that it was recognizing the VALIDATE_SCHEMA environment
variable.

Finally, I inspected the product catalog data after installation, to
make sure that the new cray-product-catalog-update image works both
for inserting the data as well as when run as a part of the
configuration import Kubernetes job, and that the 'active' flag was
set.