CRAYSAT-1289: Pull helm charts from arti instead of helmrepo
The http://helmrepo.dev.cray.com:8080/ location (i.e. Helm Chart Museum)
has been deprecated by the DST team for a while, and apparently new
chart builds are not being published to that location. In addition, it
is very slow and has caused build failures in the past. Switch to using
the helm repos in arti instead.

Note that the version of the sat-cfs-install docker image and helm chart
must be updated to a later version which was built at a later date when
DST had made the change to publish helm charts to arti instead of
helmrepo. Without this change the build fails to find the helm chart.

Test Description:
Allowed Jenkins to build release distribution, downloaded it, extracted
it, and examined the expected sat-cfs-install helm chart was present.