CRAYSAT-1316: Include the fixed version of cray-sat-podman
There was an issue where the version of the `cray/cray-sat` container
image referenced by the `cray-sat-podman` wrapper scripts was not the
correct latest master version, which resulted in a broken SAT install.
Fix it by including the fixed version of the `cray-sat-podman` RPM.

Test Description:
Built and tested install on lemondrop.