CRAYSAT-1368: Update vendor directory path
The git vendor'd release repository has been changed to come from HPE
GitHub instead of Cray Bitbucket. As a result, the path where the
vendor'd dependency is stored has changed. Update references to it in
the `release.sh` script

Test Description:
Tested that release distribution can be built (at least up to this point
in the script).