CRAYSAT-1368: Maintain existing container locations in Nexus
Modify the `release.sh` script to move the downloaded docker images from
a `docker/arti.dev.cray.com` directory into a `docker/cray` directory in
the build directory. This will result in docker images remaining in
their same paths in the container image registry in Nexus. This needs to
be done due to a change in the behavior in `skopeo-sync` which would
upload images into a path containing `arti.dev.cray.com` if that
directory exists in the `docker` dir in the release distribution build
directory.

This change is being made to avoid having to modify the repository
specified in the podman wrapper script in `cray-sat-podman` and to avoid
having to modify the repository specified in our `sat-cfs-install` helm
chart.

In the future, we may get rid of the `sat-cfs-install` helm chart
altogether and instead invoke a CSM-provided `cf-gitea-import` container
image directly in our installer with `podman`.

Test Description:
Will be tested with a batch of other changes to this repo via a SAT
product stream install on a CSM 1.2 system and a CSM 1.0 system.