CRAYSAT-1368: Specify compatible values to cray-import-config helm chart
Specify the `config_image.tag` in a way that is compatible with the new
2.0.0 version of the sat-cfs-install Helm chart which is based on the
2.0.0 version of the cray-import-config Helm chart. That version
requires the configuration import image be specified with `repository`
and `tag` keys under `config_image.image` in the Chart values.

The the 2.0.0 version of the cray-import-config Helm chart now looks for
the cray-product-catalog-update image in a repository with the
`artifactory.algol60.net` prefix in the Nexus image registry.  Since we
are bringing our own version of the `cray-product-catalog-update` image
that we are already using via `podman` in `install.sh`, and we were
specifying this version in the cray-import-config values of the loftsman
manifest, change the `repository` value under `catalog.image` to the
repository where we upload the `cray-product-catalog` image. Otherwise,
the image may fail to be found if the version we specify here is not the
version included in the CSM release.

Test Description:
Will be tested in conjunction with the new 2.0.0 version of the
sat-cfs-install Helm chart by building a release distribution containing
that version and then installing it on CSM 1.0 and 1.2 systems.