CRAYSAT-1418: Create new update-mgmt-ncn-configs.sh install script
Move the logic for updating the NCN personalization CFS configuration
from the install script, `install.sh`, to a separate script,
`update-mgmt-ncn-configs.sh` that exposes more options to the admin.

Also refactor common install utility functions into separate
`install-util.sh` script and source it in `install.sh` and
`update-mgmt-ncn-cfs-config.sh`.

Update product stream version to 2.4.2

The `cfs-config-util` image is no longer delivered into Nexus by the
installation of the SAT product. It is now packaged into the `vendor`
directory of the release distribution as a tar file. This is done for
consistency with the way `cray-nexus-setup` and `skopeo` are packaged
with the product. None of these images are used outside of the install
scripts in the release distribution, so it is not necessary to upload
them into Nexus.

Since we don't need to pull `cfs-config-util` from internal artifactory
anymore, remove the `INTERNAL_ARTI_DIR` variable definition and usages
from `release.sh`.  This kind of makes the `find` logic to find image
directories unnecessary, but it's more robust than the simple
wildcarding that was done before, so keep it.

Test Description:
Tested by building a release and installing it on baldar. Ran the new
`update-mgmt-ncn-configs.sh` script to update the management NCN CFS
configuration with various options.