CRAYSAT-1526: Update cray-sat image to 3.19.0 and declare SAT 2.4.8
Update the cray-sat image to 3.19.0 and declare a new SAT product stream
version of 2.4.8.

The main changes being pulled in here are the changes implemented under
the epic CRAYSAT-1385.

Test Description:
Build from Jenkins will be installed on hermod and tested in conjunction
with the latest version of hpc-shasta-software-recipe release
distribution for Papaya. The bootprep files delivered by that release
distribution will be tested with the new version of `sat bootprep`
delivered in this new SAT release.