CRAYSAT-1542: Fix ansible failure in fresh install
This commit adds a task to check if the SAT config file exists
prior to setting the permissions on the file. To avoid repeating
the path of the SAT config file, a new variable, "sat_config_file"
is added to defaults/main.yml.

This commit also increments the version.

Test Description:
* Installed this branch on an internal system
* I manually removed the SAT config file from ncn-m002, and then
  ran NCN personalization. Verified that the task succeeded on
  both ncn-m001 and ncn-m002.