CRAYSAT-1559: Update to latest main branch builds
Update each component to their latest builds from their main branches.
For most of these, the only reason there is a new main branch build is
because of the license checker change committed to the repos.

For the cray-sat container image, this includes a fix for CRAYSAT-1552.

Test Description:
Tested that I can build the release distribution with `release.sh`.