CRAYSAT-1614: Add an IUF Product Manifest file to SAT
Add an IUF Product Manifest file to the SAT release distribution that
defines the behavior of the SAT installation within the Install and
Upgrade Framework (IUF) that is going to be delivered in a subsequent
CSM release.

Also validate it using a new function that will be added to
hpc-shastarelm-release when that gets updated.

Update .gitignore to ignore generated IUF Product Manifest.

Test Description:
Built locally and confirmed that the file `iuf-product-manifest.yaml` is
included in the resulting release distribution.