CRAYSAT-1614: Fix bug in substitution of RELEASE_VERSION variable
The `RELEASE_VERSION` environment variable is used by
`render_templates.sh`, but it is not available there because it is not
exported. Change `release.sh` to source the script instead of running it
in a separate process to avoid issues like this.

Also switch the order so that this script comes after the `requires`
call that looks for the presence of `sed`, which is used by this script.

This hasn't been an issue with our official remote builds on Jenkins
because the Jenkinsfile sets the `RELEASE_VERSION` variable with the
output of `version.sh` separately.

Test Description:
Built the release distribution locally.