CRAYSAT-1601: Ensure credentials are available for RPM
This commit modifies the Jenkins build process for the SAT product
stream so that sat-rpms can be made private.

This is based on PR#1654 to Cray-HPE/csm. Note that fewer changes
are required here because SAT is using the hpc-shastarelm-release
library when running rpm-sync, which appears to already have support
for credentials.

This commit also creates a temporary file containing repository
credentials, which is needed by the rpm-sync function. This approach
is based on an approach found in the CSM release script.

It ensures that the path to the file is exported as the environment
variables REPOCREDSPATH and REPOCREDSFILENAME, variables which are
expected to be set by the release.sh script found in
hpc-shastarelm-release.

This commit also increments the version number.