CRAYSAT-1601: Ensure credentials are available for RPM
This commit modifies the Jenkins build process for the SAT product
stream so that sat-rpms can be made private.

This is based on PR#1654 to Cray-HPE/csm. Note that fewer changes
are required here because SAT is using the hpc-shastarelm-release
library when running rpm-sync, which appears to already have support
for credentials.

This commit also increments the version number.

Test Description: Verified I was able to build the product stream.