CRAYSAT-1751: Update cray-sat version to 3.21.7
Update the cray-sat container version to 3.21.7. This fixes several CVEs
in Python packages. It also pulled in several CVE fixes in apk packages
included in the cray-sat container.

Also pull in the latest sat-install-utility container which has fixed
several CVEs in Python packages.

Also update to the latest docs build.