CRAYSAT-1766: Update cray-sat version to 3.25.5
Update the cray-sat container version to 3.25.5. This fixes security
vulnerabilities in the urllib3 and cryptography python packages. It also
pulls in several CVE fixes in apk packages included in the cray-sat
container.

It also pulls in a build fix and a `sat showrev` and `sat setrev` bug
fix.

Test Description:
This container image version was tested on mug.